+++ parallel/ParallelSolver.cc
@@ -469,7 +469,7 @@ lbool ParallelSolver::solve_(bool do_simp, bool turn_off_simp) {
     if (firstToFinish && status == l_True) {
         extendModel();
         // Extend & copy model:
-        if (model.size() < nVars()) model.resize(nVars());
+        if ((int)model.size() < nVars()) model.resize(nVars());
         for (int i = 0; i < nVars(); i++) model[i] = value(i);
     } else if (status == l_False && conflict.size() == 0)
         ok = false;
